refactor: move get version api to info route group
