+++ internal/node/component/network/handler_network.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 
 	"github.com/labstack/echo/v4"
-	"github.com/rss3-network/node/internal/constant"
 	"github.com/rss3-network/node/schema/worker"
 	"github.com/rss3-network/protocol-go/schema/network"
 )
@@ -33,10 +32,6 @@ type WorkerConfigResponse struct {
 
 type EndpointConfigResponse struct {
 	Data Endpoint `json:"data"`
-}
-
-type VersionResponse struct {
-	Data Version `json:"data"`
 }
 
 // GetNetworksHandler returns the list of all networks that the node can support.
@@ -129,17 +124,5 @@ func (c *Component) GetEndpointConfig(ctx echo.Context) error {
 
 	return ctx.JSON(http.StatusOK, EndpointConfigResponse{
 		Data: config,
-	})
-}
-
-// GetVersion returns the version of the network component.
-func (c *Component) GetVersion(ctx echo.Context) error {
-	tag, commit := constant.BuildVersionDetail()
-
-	return ctx.JSON(http.StatusOK, VersionResponse{
-		Data: Version{
-			Tag:    tag,
-			Commit: commit,
-		},
 	})
 }
